fix notify, launch Count query.

+++ app/src/main/java/com/chs/yourapphistory/data/paging/GetPagingWeekAppNotifyInfo.kt
@@ -49,7 +49,7 @@ class GetPagingWeekAppNotifyInfo(
                 ).map {
                     it.key.toLocalDate().dayOfWeek
                         .getDisplayName(TextStyle.SHORT, Locale.KOREAN) to
-                            it.value.toInt()
+                            it.value
                 }
             }
